refactor(nvim): use a split buffer for viewing codecompanion diffs
I prefer to view diffs in a split buffer than a floating window, because
then I can also switch to other buffers and explore the codebase. The
floating window is a modal window that captures focus and makes it more
difficult to see other buffers, including the code companion chat, which
has extra information about the diff.